// src/backend/services/files/interfaces/getFiles.ts
import { FitEnum } from 'sharp';

/* ****************************************************************************************************************** */
interface ImageParams {
  noColor?: boolean;
  width?: number;
  height?: number;
  quality: number;
  fit?: keyof FitEnum;
  withoutEnlargement?: boolean;
}

/* ****************************************************************************************************************** */
interface FileWhere {
  id?: string;
  num?: number;
}

/* ****************************************************************************************************************** */
interface FileOutput {
  content: Buffer;
  contentType?: string;
  name?: string;
}

/* ****************************************************************************************************************** */
export {
  ImageParams,
  FileWhere,
  FileOutput,
};

/* ****************************************************************************************************************** */